dnt_quant_lab/backend/requirements.txt
fastapi
uvicorn
pandas
numpy
plotly
requests
google-generativeai
python-dotenv
vnstock3
requests-cache
